feat(css): split postcss printer into separate file (#1969)
